Damage is now handled by each entity
-Damage script deleted
