+++ Assets/Scripts/DamageZone.cs
@@ -1,0 +1,23 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class DamageZone : MonoBehaviour
+{
+
+    public int damage = 2;
+    public float range = 4f;
+    public Health entityHealth;
+
+    // Start is called before the first frame update
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+}
