+++ Assets/Scripts/Weapon.cs
@@ -9,11 +9,13 @@ public class Weapon : MonoBehaviour
     public float attackSpeed = 1f; //Seconds between each shot
     public float range = 5f;  //Range before destruction;
     private  Vector3 startPosition;
+    public Health entityHealth;
+
     // Start is called before the first frame update
     void Start()
     {
         startPosition = this.transform.position;
-        GetComponent<Damage>().damage = damage;
+        //GetComponent<Damage>().damage = damage;
     }
 
     // Update is called once per frame
@@ -23,6 +25,22 @@ public class Weapon : MonoBehaviour
 
         if(Mathf.Abs(Vector3.Distance(startPosition, this.transform.position)) > range){
             Destroy(this.gameObject);
+        }
+    }
+
+    private void OnCollisionEnter(Collision collision){
+
+        if(entityHealth == null){
+
+            entityHealth = collision.gameObject.GetComponent<Health>();
+            if(entityHealth == null)
+                return;
+        }
+
+        if(collision.gameObject.tag == "Enemy"){
+
+            Destroy(gameObject);
+            entityHealth.takeDamage(damage);
         }
     }
